some refactorin - split open/auth routes, use hono builtins for auth/logging
