+++ src/api/routes/open/index.ts
@@ -1,0 +1,6 @@
+import { Hono } from "hono";
+import login from "./login";
+
+const openRoutes = new Hono().route("/", login);
+
+export default openRoutes;
