+++ src/server.tsx
@@ -1,71 +1,90 @@
-import React from "react";
+/** @jsxImportSource hono/jsx */
 import { Hono } from "hono";
-import { serveStatic } from "@hono/node-server/serve-static";
-import { secureHeaders } from "hono/secure-headers";
-import { csrf } from "hono/csrf";
-import { renderToString } from "react-dom/server";
-import { routes } from "./api";
-import { authMiddleware } from "./api/middleware/auth";
 import { serve } from "@hono/node-server";
+import { serveStatic } from "@hono/node-server/serve-static";
+import { csrf } from "hono/csrf";
 import { cors } from "hono/cors";
-import { env } from "./env";
-const server = new Hono({});
+import { logger } from "hono/logger";
+import { jwt } from "hono/jwt";
+import { jsxRenderer } from "hono/jsx-renderer";
+import { secureHeaders } from "hono/secure-headers";
+import { rateLimiter } from "hono-rate-limiter";
 
-// Replace the existing CORS and CSRF configuration with:
-const protocol = process.env.NODE_ENV === "production" ? "https" : "http";
-const port = process.env.NODE_ENV === "production" ? "" : `:${env.CLIENT_PORT}`;
+import { authRoutes, openRoutes } from "./api/routes";
+import { env } from "./env";
+
+const protocol = env.NODE_ENV === "production" ? "https" : "http";
+const port = env.NODE_ENV === "production" ? "" : `:${env.CLIENT_PORT}`;
 const origin = `${protocol}://${env.DOMAIN}${port}`;
 
-server.use(
-	"*",
-	cors({
-		origin: [origin],
-		credentials: true,
-		allowMethods: ["GET", "POST", "PUT", "DELETE", "OPTIONS"],
-		allowHeaders: ["Content-Type", "Authorization", "X-CSRF-Token"],
-	})
-);
+const server = new Hono()
+	// CORS
+	.use(
+		"*",
+		cors({
+			origin: [origin],
+			credentials: true,
+			allowMethods: ["GET", "POST", "PUT", "DELETE", "OPTIONS"],
+			allowHeaders: ["Content-Type", "Authorization", "X-CSRF-Token"],
+		})
+	)
 
-server.use(
-	"*",
-	csrf({
-		origin: (requestOrigin) => {
-			const allowedOrigins = [origin];
-			return allowedOrigins.includes(requestOrigin);
-		},
-	})
-);
+	// CSRF
+	.use(
+		"*",
+		csrf({
+			origin: (requestOrigin) => {
+				const allowedOrigins = [origin];
+				return allowedOrigins.includes(requestOrigin);
+			},
+		})
+	)
 
-server.use(secureHeaders());
+	// Secure headers
+	.use(secureHeaders())
 
-// API routes
+	// Logger
+	.use(logger())
 
-// Add your routes before the catch-all middleware
-server.route("/api", routes);
+	// Non-authenticated routes
+	.use(
+		"/api/open",
+		rateLimiter({
+			windowMs: 1 * 60 * 1000, // 1 minute
+			limit: 5,
+			standardHeaders: "draft-6",
+			keyGenerator: () => env.RATELIMIT_KEY,
+		})
+	)
+	.route("/api/open", openRoutes)
 
-// Then add the auth middleware only for protected routes
-server.use("/api/*", authMiddleware);
-// Public assets
-server.use("/public/*", serveStatic({ root: "./public" }));
-server.use("/assets/*", serveStatic({ root: "./dist/assets" }));
+	// Authenticated routes
+	.use(jwt({ secret: env.JWT_SECRET, cookie: "auth_token" }))
+	.route("/api", authRoutes)
 
-// Serve HTML for all other routes
-server.get("/*", (c) => {
-	const html = renderToString(
-		<html lang="en">
-			<head>
-				<meta charSet="utf-8" />
-				<meta name="viewport" content="width=device-width, initial-scale=1" />
-				<title>Thamyris</title>
-			</head>
-			<body>
-				<div id="root"></div>
-				<script type="module" src="/src/client.tsx"></script>
-			</body>
-		</html>
+	// Public assets
+	.use("/public/*", serveStatic({ root: "./public" }))
+	.use("/assets/*", serveStatic({ root: "./assets" }))
+
+	// Serve HTML for all other routes
+	.get(
+		"/*",
+		jsxRenderer(() => (
+			<html lang="en">
+				<head>
+					<meta charSet="utf-8" />
+					<meta name="viewport" content="width=device-width, initial-scale=1" />
+					<title>Thamyris</title>
+				</head>
+				<body>
+					<div id="root"></div>
+					<script type="module" src="/src/client.tsx"></script>
+				</body>
+			</html>
+		))
 	);
-	return c.html(html);
-});
+// TODO: Compress response
+// .use(compress());
 
 // Only start the server if we're in the server process
 serve({
