+++ src/utils/api.ts
@@ -1,4 +1,4 @@
-import { ApiRouteType } from "@/api";
 import { hc } from "hono/client";
+import { ApiRouteType } from "@/api/routes";
 
 export const api = hc<ApiRouteType>("/api");
